Moved SpectrogramSequence to its own module spectrogram.py in anticipation that this module may be moved out of IRISpy and into SunPy due to its generalized nature.
